SWORD25: Rename double serialization file to better represent what it is
AKA functions, rather than a class
