+++ engines/sword25/util/double_serialization.h
